Move vault editing into dialog
