+++ source/setup/index.js
@@ -6,12 +6,11 @@ import { ConnectedRouter } from "react-router-redux";
 import store from "./redux/index.js";
 import history from "./redux/history.js";
 import AddArchivePage from "./containers/AddArchivePage.js";
-import ArchiveUnlockPage from "./containers/ArchiveUnlockPage.js";
+import VaultPage from "./containers/VaultPage.js";
 import LoadingModal from "./containers/LoadingModal.js";
 import SaveCredentialsPage from "./containers/SaveCredentialsPage.js";
 import AboutPage from "./components/AboutPage.js";
 import UnlockAllArchivesPage from "./containers/UnlockAllArchivesPage.js";
-import VaultPage from "./containers/VaultPage.js";
 import App from "../shared/containers/App.js";
 import { trackMouseMovement } from "../shared/library/mouseEvents.js";
 import { trackKeydownEvent } from "../shared/library/keyboardEvents.js";
@@ -26,11 +25,10 @@ ReactDOM.render(
         <ConnectedRouter history={history}>
             <App>
                 <Route path="/add-archive" component={AddArchivePage} />
-                <Route path="/access-archive/:id/:state" component={ArchiveUnlockPage} />
+                <Route path="/access-archive/:id/:state" component={VaultPage} />
                 <Route path="/save-new-credentials" component={SaveCredentialsPage} />
                 <Route path="/about" component={AboutPage} />
                 <Route path="/unlock" component={UnlockAllArchivesPage} />
-                <Route path="/vault/:id" component={VaultPage} />
                 <LoadingModal />
             </App>
         </ConnectedRouter>
